Move health.go to tangle plugin
